Added project structure folder for MVC
