+++ Movs/Movs/Presentation/Controller/ViewController.swift
@@ -15,6 +15,4 @@ class ViewController: UIViewController {
         // Do any additional setup after loading the view, typically from a nib.
     }
 
-
 }
-
