+++ Movs/Movs/Supporting Files/AppDelegate.swift
@@ -1,0 +1,73 @@
+//
+//  AppDelegate.swift
+//  Movs
+//
+//  Created by Adann Simões on 12/11/18.
+//  Copyright © 2018 SAS. All rights reserved.
+//
+
+import UIKit
+import CoreData
+
+@UIApplicationMain
+class AppDelegate: UIResponder, UIApplicationDelegate {
+
+    var window: UIWindow?
+
+
+    func application(_ application: UIApplication,
+                     didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+        return true
+    }
+
+    func applicationWillResignActive(_ application: UIApplication) {
+        
+    }
+
+    func applicationDidEnterBackground(_ application: UIApplication) {
+        
+    }
+
+    func applicationWillEnterForeground(_ application: UIApplication) {
+        
+    }
+
+    func applicationDidBecomeActive(_ application: UIApplication) {
+
+    }
+
+    func applicationWillTerminate(_ application: UIApplication) {
+
+        self.saveContext()
+    }
+
+    // MARK: - Core Data stack
+
+    lazy var persistentContainer: NSPersistentContainer = {
+
+        let container = NSPersistentContainer(name: "Movs")
+        // storeDescription
+        container.loadPersistentStores(completionHandler: { (_, error) in
+            if let error = error as NSError? {
+
+                fatalError("Unresolved error \(error), \(error.userInfo)")
+            }
+        })
+        return container
+    }()
+
+    // MARK: - Core Data Saving support
+
+    func saveContext () {
+        let context = persistentContainer.viewContext
+        if context.hasChanges {
+            do {
+                try context.save()
+            } catch {
+                let nserror = error as NSError
+                fatalError("Unresolved error \(nserror), \(nserror.userInfo)")
+            }
+        }
+    }
+
+}
